requirements.txt
fastapi>=0.135,<0.136
uvicorn>=0.42,<0.43
python-multipart>=0.0.22,<0.0.23
rank-bm25>=0.2.2,<0.3
openai>=2.29,<3
chromadb>=1.5,<1.6
transformers>=5.3,<5.4
torch>=2.11,<2.12
torchvision>=0.26,<0.27
numpy>=2.4,<2.5
pillow>=12.1,<13
accelerate>=1.12,<2
sentencepiece>=0.2,<0.3
pymupdf>=1.26,<1.27
pandas>=2.3,<2.4
openpyxl>=3.1,<3.2
ultralytics>=8.4,<8.5
